refactor: extract DiagnosticsLogLevel.swift from DiagnosticsTypes.swift
Splits the `DiagnosticsLogLevel` enum (with its OSLog `osLogType`
bridge) out of DiagnosticsTypes.swift so that file holds only
DiagnosticsLogCategory and the shared event-name types.

Byte-identical move of lines 14-36. Both files import OSLog + Foundation.

DiagnosticsTypes.swift: 92 → 68 lines. Tests: 667.
